Change Flow implementation to not cast to readonly
This makes get work with non-object types, like arrays, and some more
complex types like unions and $Shape<T>. As a tradeoff, get doesn't work
with classes or interfaces now.

I originally decided that the opposite end of the tradeoff would be
better, but then ran into lots of issues with complex type unions. At the
end of the day, I think usage with plain types instead of classes might be
more common. If this turns out not to be the case, we can always add
another function with different type signature.